src/data/reddit/requirements.txt
python-dotenv
PyYAML
praw
neo4j
neomodel